requirements.txt
alembic==1.16.2
annotated-types==0.7.0
ansicolors==1.1.8
anyio==4.9.0
app==0.0.1
asyncpg==0.30.0
certifi==2025.6.15
click==8.2.1
colorama==0.4.6
dish==0.0.0
dulwich==0.22.8
fastapi==0.115.13
functions==0.7.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.6.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
pytest==8.4.1
python-dotenv==1.1.0
PyYAML==6.0.2
schemas==0.7.1
services==0.1.1
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.34.3
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
